+++ .storybook/webpack.config.js
@@ -4,25 +4,31 @@ module.exports = (baseConfig, env, config) => {
   config.module.rules.push({
     test: /\.(ts|tsx)$/,
     use: [{
-      loader: require.resolve('awesome-typescript-loader')
-    }, {
-      loader: require.resolve('react-docgen-typescript-loader')
-    }, {
-      loader: require.resolve('ts-loader')
-    }, {
+    //   loader: require.resolve('awesome-typescript-loader')
+    // }, {
+    //   loader: require.resolve('react-docgen-typescript-loader')
+    // }, {
+    //   loader: require.resolve('ts-loader')
+    // }, {
       loader: require.resolve('babel-loader'),
-      options: {
-        presets: [['react-app', { flow: false, typescript: true }]],
-      }
-    }, {
-      loader: require.resolve('html-loader')
-    }]
+    // }, {
+    //   loader: require.resolve('html-loader')
+    // }],
+      // options: {
+      //   presets: [['react-app', { flow: false, typescript: true }]],
+      // },
+      loader: require.resolve('ts-loader'),
+      // loader: require.resolve("css-loader"),
+      // options: {
+      //   modules: true,
+      // }
+    }],
   });
 
   config.module.rules = config.module.rules
     .filter(r => r.test.toString() !== "/\\.css$/")
     .concat({
-      test: /\.css$/,
+      test: /\.(css|scss)$/,
       use: [
         {
           loader: require.resolve("style-loader"),
